requirements.txt
cycler==0.10.0
matplotlib==1.5.1
numpy==1.11.0
Pillow==3.2.0
pyparsing==2.1.1
python-dateutil==2.5.2
pytz==2016.3
scikit-learn==0.17.1
scipy==0.11.0
six==1.10.0
